requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
openpyxl==3.1.5
rapidfuzz==3.12.2
Pillow==12.0.0
requests==2.32.3
pytesseract==0.3.13
